Updated code after new data
Import data, recevied 16-Sep-2019 from Sarah Almond, with data processing errors fixed.

1. Patient 7113: if age = 0 then set to missing or is it an infant?

Patient was born and died in the hospital.

2. Date of Admission after 1950… can you please check these patient IDs? All Fixed.

4825

3902

3880

750

334

3. Patient 479 - Admit date is Feb 29, 1862, but that was not a leap year. Should this be set to Feb 28 or a different year? 1860 and 1864 were nearest leap years. I double checked the entry and it very clearly says 2/29/1862. Manual change in code by Nab to 3/1/1862 for data cleaning.

4. Date of discharge

Patient 2701  is "18996-11-30” should it be "1896-11-30”? Fixed.

Patient 3903 is "18904-11-30” should it be "1894-11-30”?  Fixed.

5. One patient (7152) had gender listed as “M;F” and no mention of sexual deviancy. Is this a data entry error? 7152 is marked as both in the admissions ledger (without comment) but is identified as “Male” in the General Case Book. I have changed gender to M.

6. Patient 1694 seems like it might be disentangled into 2 rows, each representing an admission with the previousadmisions incrementing up one

dateofdischarge:    1890-10-01/1891-03-17

lengthofstay:   _ _ _ 5y;7m;10d

dateofadmission: 1885-08-07

remarksfreetext: and discharged Mar. 17, 1891 by order of Ex. Com.

We have a couple of these, where the patient is initially released and comes right back to the hospital for a longer stay, only to be released again about 6 months later. I usually double them up (two instances of the same number), but hadn’t caught this one. I’ve fixed it in the spreadsheet.